+++ js/util.js
@@ -4,3 +4,4 @@ function getRandomInteger (min, max) {
   const result = Math.random() * (upper - lower + 1) + lower;
   return Math.floor(result);
 }
+export {getRandomInteger};
